requirements.txt
fastapi==0.109.0
python-multipart>=0.0.9
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
pyrogram==2.0.106
tgcrypto==1.2.5
